7.4

+++ controller/videoController.js
@@ -69,7 +69,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -92,8 +96,13 @@ export const deleteVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
+  const video = await Video.findById(id);
   try {
-    await Video.findOneAndRemove(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove(id);
+    }
   } catch (error) {
     console.log(error);
   }
